TKSS-420: Test with different JDKs
